[CL-2333] Refactored analytics chart component into separate components
