+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/VisitorsTrafficSourcesWidget/VisitorTrafficSourcesCard.tsx
@@ -7,7 +7,7 @@ import useNarrow from 'containers/Admin/reporting/hooks/useNarrow';
 // components
 import { Box, Title } from '@citizenlab/cl2-component-library';
 import PieChart from 'components/admin/Graphs/PieChart';
-import NoChartData from '../AnalyticsChartWidget/NoChartData';
+import NoChartData from '../NoChartData';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
