+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/typings.ts
@@ -1,0 +1,6 @@
+export interface ChartWidgetProps {
+  title: string;
+  projectId: string | undefined;
+  startAt: string | null | undefined;
+  endAt: string | null;
+}
